update 3rd libraries

+++ jxdribbble/jxdribbble_DebutsViewController.m
@@ -137,10 +137,11 @@
     {
         url = [NSURL URLWithString:shot.image_url];
     }
+
     [cell.shot_imageView setImageWithURL:url
                         placeholderImage:[UIImage imageNamed:@"placeholder.png"]
                                  options:SDWebImageLowPriority
-                                progress:^(NSUInteger receivedSize, long long expectedSize) {
+                                progress:^(NSInteger receivedSize, NSInteger expectedSize) {
                                     double p = (double)receivedSize/(double)expectedSize;
                                     weakCell.hud.progress = p;
                                 } completed:^(UIImage *image, NSError *error, SDImageCacheType cacheType) {
